Build Gimp 2.99
- jhbuildrc: Gimp 2.99.x requires python3
- package: Create build files for 2.99 and python3

Creates build script, bundle and plist template for 2.99

Changes the inbuilt python to 3.

- .circleci: build gimp299.sh instead of gimp.sh.

Modify the many parts of the circleci config to make it ready to build
gimp 2.99

- Update xcode version
- install python 3.9.7
- .circleci: Comment out webkit

Webkit is failing to build so this comments it out until we have it
building again and against gtk3

- .circleci: Correctly set $PYTHON variable
- jhbuildrc: Set PYTHON env variable correctly.

Cf. https://gitlab.gnome.org/Infrastructure/gimp-macos-build/-/merge_requests/31#note_1279206

- glib-networking is a webkit dependency for libsoup
- package: Resolve missing icon and plugin @rpath

Installs the Adawaita and hicolor icons in two places in the
app tree because otherwise we get missing icon errors.

Also fixes the @rpaths for plugins.

Discussion about the missing icon here:
https://gitlab.gnome.org/GNOME/gimp/-/issues/7319#note_1283260

- 2.99 bundle
- Apparently unneeded libpyglib
- gimp299.bundle: Fix plugin loading

Unfortunately, because there is no support for subdirectories in the
binary tag in `mac-gtk-bundler`, each plugin has to be listed
separately.
NOTE: If plugins are added, they are likely to be forgotten.

- .cricleci: Fix notarize script to use DMG name
- Upgrade create-dmg to version 1.0.9
- .circleci: Never fail signing and notarizing TODO

Other accounts than Infrastructure don't have the right secrets
on circle ci to have this succeed. A better approach would be to
detect that we are on the right account and otherwise avoid
calling the scripts.

However, it wasn't immediately clear how to frame the conditional
so it was skipped.

- .circleci: Create new cache key for gimp 2.99

- jhbuildrc: Point temporarily to lukaso/gtk-osx

Start process of build 2.99 on Mac and point to temp repo while waiting
for https://gitlab.gnome.org/samm-git/gtk-osx/-/merge_requests/10
to be merged.

- applescript: get permissions right